requirements.txt
streamlit
python-dotenv
google-generativeai
textblob
bycrypt